applets/notes: make default paste action enable clickable URLs by default
Currently, when pasting text into the Notes widget, formatting is removed, and as
a result URLs are not clickable. When using notes to store a link for later usage,
this always requires the user to manually select the link text and copy and paste
it into a browser, which is annoying. This commit solves that issue by making the
default paste action only strip non-link formatting; links remain formatted and
hence clickable.