Don't show Plasma Calculator widget in search results
This widget looks quite app-like when open in a windowed form using the
"Windowed widgets" KRunner runner search result. When launched in that
way, it can give the user the impression that it's a real app. This
presents some issues:

1. If the user has a real calculator app installed, two calculator
results will appear in search results, which is weird and confusing.
People will wonder, "Why does KDE ship with two calculator apps"?

2. If the user has no calculator apps installed, this will appear in
search results anyway and they will think it is the official KDE
calculator app and use it instead of using the "download a calculator
app" search result from the AppStream runner. The widget is quite simple
with almost no features, which is fine for a widget. But if the user
misunderstands this and thinks it's the official KDE calculator app,
they'll be lacking those features and form an incorrect impression of
KDE's ability to write good calculator apps, because in truth, we have
two powerful high quality calculator apps (KCalc and Kalk).

See https://bugs.kde.org/show_bug.cgi?id=463809.