applets/katesessions: set Plasmoid.icon to a symbolic icon when in panel
Previously it wasn't set at all, so it was just falling back to the
generic icon.

Currently this code produces no changes since there is no
"kate-symbolic" icon in the Breeze icon theme, and therefore presumably
not in any other ones either. But, once one exists, it will be used here
automatically.

CCBUG: 473278