applets/weather: Fix SwitchPanel dynamics
The QtQuick Container API changed since QtQuick.Controls 2.3
(Qt 5.10) deprecating `removeItem(index)`.

This can cause some issues when the pages are added and then
removed, for instance when some warnings appear, or by changing
the station to a different provider.

To fix this we need to refactor SwitchPanel into a model-based
Repeater structure that dynamically adds/removes the items
depending on the conditions (making items non-`visible` do not
work because they still occupy the space)

BUG: 472171
FIXED-IN: 5.91.90