applets/weather: Fix spacing on forecast view
Remove a now unncessary size hint, which was artificially
increasing the minimum icon separation in the forecast view,
allowing for a more compact view.

This was specially wrong for the Environment Canada provider,
where we hide the "night" labels, but were taking their width
into account.

The Day/Night labels are also now left-aligned.