cube: Fix error handling of loaded QML
Delegate takes a Component as an argument, it acts as a loader. By
passing a loader to the delegate we effectively make a loader that loads
other items.

More importantly this breaks the error detection at a kwin level, we can
only tell that the loader was created successfully not that the loader
loaded something successfully. This leaves the user with a black screen
in the case of a missing dependency.