applets/weather: Move temperature label to the other side of the icon
In left-to-right languages, people read UIs from left to right, which
means the most important information should be on the left. The
temperature is clearly important because we also make it huge and bold.
Let's acknowledge this visual importance by using layout position too.

This layout reverses correctly when using an RTL language.

| Before | After |
| ------ | ----- |
| ![Before](/uploads/dd4fc7dcbeda868d3ad622664b0eb53a/Before.jpeg) | ![After](/uploads/4580d91a8054a81f95b18d05e73ed999/After.jpeg) |

@teams/vdg  @iasensio