applets/weather: Reduce forecast layout minimum width
When the forecast has a night row, we show "Day" and "Night" labels
as the row headings. If the forecast starts at night, we can play
with the "Night" label positioning to save some horizontal space

Technically, instead of pushing fake items (placeholders) into the
model and use a spacial case on the usual delegates, add the Day/Night
labels directly to the grid. This also simplifies the delegates.

Also reduce the grid horizontal spacing to 0, as the icon margins are
enough to compensate for this

BUG: 492243
FIXED-IN: 6.2