wallpapers/potd: reorder options
The Picture of the Day wallpaper type currently puts the Provider as the
first option. This is not unreasonable, as it is certainly an important
setting, but it breaks consistency with the Image and Slideshow types,
which both have Positioning as the first option, and puts a large gap
between the provider combobox and the thumbnail of today's picture.

This change moves the Positioning and Background color options over the
provider, increasing consistency between wallpaper types and reducing
the visual distance between the provider and the thumbnail.