applets/weather: Fix apperance and size for the umbrella emoji
If the regular font and the emoji font don't match, the height of
the bounding boxes will be different for labels with or without the
umbrella, disrupting the layout.

Let's set a fixed height using the metrics of the regular font, and
also remove the unicode marker for text-mode emojis, because it's
non-deterministic, not getting translated, and also looks worse.

BUG: 492340
FIXED-IN: 6.2