runners/datetime: Fix autotest sporadic failures
The first test compares the current's time with the result of the
runner's current time, with a precision of seconds.
Even if only milliseconds pass between the two calls, with a bad time
alignment, the "second's digit" might advance making the test fail.

Let's compare it also with the current time +1s to fix it

The other test don't have this precision problem