applets/diskquota: fix model reference
The placeholder in the disk quota applet accesses the model under the
property name model, but it is not accessible under that name outside
the listView, so an error message is printed every time the applet runs.
The placeholder also doesn't actually do anything with the value that it
generates from the model, as the resulting property isn't used anywhere.

This change removes the line that accesses the model completely.