applets/diskquota,katesessions: fix action removal
The diskquota and katesessions applets have no configuration options,
and so remove the standard action from the applet context menu. To do
this, however, they use Plasma 4 API (removeAction) that doesn't work
anymore and instead prints an error message.

This change ports this to the new API (removeInternalAction), silencing
the error messages and actually removing the configure option from the
context menu.