weather,fuzzyclock: don't set invalid point size
The weather and fuzzyclock applets set both point size and pixel size
for the label font, and the point size is set to an invalid value – the
documentation is clear that the needs to be greater than zero. This
causes an error message to be printed. The comments seem to indicate
that this is necessary or the Text.fit fontSizeMode would break, but
this seems rather old and in my testing Text.fit seems to work just the
same without it.

This change removes this property assignment.