applets/comic: enforce minimum FullViewWidget size
The comic applet contains a prominent button that shows the current
comic full size in a popup window. In case there is no image (and no
fallback), the opened pop-up has a size of 0, and the applet goes into a
permanent loop trying to handle this. If the applet is running in
plasmashell, this will hang plasmashell completely.

This change enforces a minimum popup size of 4×gridUnit on each side;
this is enough to show a visible window, but should still be smaller
than even a small realistic comic.