kwin/effects: Mention the cube effect's need for manual activation
The Cube effect requires a two-step process to activate: first enable
it in the KCM, and activate it using a keyboard shortcut — either the
default one or a custom one. But this may not be obvious to users,
who can think that simply enabling it in the KCM is enough.

To improve this, mention the need for manual activation using a keyboard
shortcut.

CCBUG: 492936