runners/converter: Fix thread affinity of currency update timer
Amends 8f164924d06015db13a8b703fdfa1258d7f8d718.

The runner lives in a separate thread, but the currency update timer
lives on the main thread. Furthermore, the update timer is owned by
the runner, which is not okay because its thread affinity is different.
It also means that the event loop on the main thread may not properly
clean up its timer list.

In order to fix the issue, this change makes the update timer a child
of the runner object. With that, both will have the same thread affinity.

BUG: 488713


(cherry picked from commit 1245968618afff3a0bba5e2c6088e77f0c9f8aff)

Co-authored-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>