applet/webbrowser: add persistent storage
This adds persistent storage to the webbrowser applet.

Sometimes you want to be able to store a state of an website like e.g. login data.

This creates a webengine browser profile stored in `~/.local/share/plasmashell/QtWebEngine/org.kde.plasma.webbrowser/` and introduces the option to turn [private browsing](https://doc.qt.io/qt-6/qml-qtwebengine-webengineprofile.html#offTheRecord-prop) on and off (defaults to `on` like the behaviour in the default profile).

@fusionfuture