applets/userswitcher: fix showing other logged-in users' avatars
`model.icon` is returned as `/var/lib/AccountsService/icons/account1`
if an image is defined, but this is being prefixed with `qrc:`. 

Prefixing the path with `file://` solves the issue, checked in the
Kirigami channel on Matrix and it seems concatenating a string is
the way to go here; we do the same thing in the lock screen at LockScreenUi.qml#L231.