applets/userswitcher: fix showing other logged-in users' avatars
`model.icon` is returned as `/var/lib/AccountsService/icons/account1`
if an image is defined, but this is being prefixed with `qrc:`. 

Prefixing the path with `file://` solves the issue, checked in the
Kirigami channel on Matrix and it seems concatenating a string is
the way to go here; we do the same thing in the lock screen at LockScreenUi.qml#L231.


(cherry picked from commit 31bf15c89187094639105c3b8c31fcc171469c98)

38d75212 applets/userswitcher: Catch avatar image resources and return with "file://"

Co-authored-by: Blazer Silving <breakingspell@gmail.com>