applets/notes: don't steal shortcuts to FloderView
Shortcuts are in the context of the window, so the notes applet will
steal any shortcut that folderview would register as well, such as
the paste shortcut.
If there is a note on the desktop, after copying a file from dolphin,
ctrl+v with folderview focused will paste the file url in the note.
as well other shortcuts such as undo/redo will have effect even if
the note is not focused

This enables the shortcut only when the context menu is open, as the same
actions are triggered as normal key events when the text field is focused

CCBUG:498867