applets/userswitcher: hide menu when switching user
When using the userswitcher applet to start a new session or switch to
an existing user, the menu remains open even after switching back to the
current user. This is almost never useful; if the user wanted to
immediately switch again, lock the screen, or reboot/shutdown, they
could have done so before switching back.

This change hides the menu after a user switch is initiated from the
userswitcher menu. For locking the screen or showing the logout screen,
this is not necessary as the menu is automatically closed by these
actions.