applets/userswitcher: set AvatarButton description
The userswitcher applet menu allows the user to access the users kcm
when clicking the Avatar. The tooltip for this is not helpful (it is
just the user name again) and there is no Accessible.description for
screen reader users either.

This change sets "Open user configuration" as the tooltip and
Accessible.description.