applets/userswitcher: don't highlight user on press
The user list contais the currently logged in user as the first entry.
There is no action associated with this entry, as it is largely
informative only, and so the hover highlight is turned off to indicate
that it is not clickable. However, it is still an AbstractButton, and
if the user clicks it there is a highlight from being pressed. As such,
it looks to the user as if it would do something, but it does not and is
not supposed to.

This change turns off the pressed highlight, making the entry appear
less clickable even if the user clicks it.