applets/katesessions: remove unnecessary MouseArea
Katesessions wraps its ItemDelegate in a MouseArea. This is not
necessary, as ItemDelegate is an AbstractButton and already contains
a MouseArea; the missing onEntered handler can be implemented with
onHoveredChanged.

This change removes this MouseArea, reducing the complexity of the code
a little by removing a level of indentation. As this requires whitespace
change anyway, the extra indentation on some of the later sections of
the file is also cleaned up.