applets/katesessions: fix keyboard activation
Pressing enter/return on the applet currently does not work at all - the
applet tries to get the info from the model using get(), but
KSortFilterProxy does not support that.

This change makes it just clicked() the delegate instead, which already
has the neccesary information. Along the way, the parameter injection
is made explicit to silence the log spam.