app/katesessions: rework keyboard handling/a11y
The applet has almost all its keyboard handling in a large onPressed
handler at the root item. This works, but is rather unwieldy and blocks
screen readers from working - the base ListView handling actually passes
focus to the selected item which informs screen readers, while the
current handling does not. Some instances of the keyboard handling
are also non-functional, such as activating items by pressing enter or
return.

This change moves most key handling locally into the components and
leverages their built-in handling more. More specific key handlers and
NeyNavigation are used where it makes sense. The general keyboard
behavior is tweaked a bit (such allowing to activate the first entry by
pressing enter without needing to arrow-down first) and non-functioning
parts are fixed. In addition, the conditions when the selected item is
reset are tweaked so that e.g. the selection is reset on closing the
menu, and small a11y tweaks are made (in particular, the number of
entries of sessions is read out).