applets/weather: Fix layout for large station names
The layout minimum width was unconditionally dependent of the station
name, which in some extreme cases, could overflow the strict width
set by the system tray.

Put a maximum cap to this minimum width and elide the label if
necessary.

This has benefits also on the other form factors (desktop and free
in the panel), as one can now reduce the applet's width even for
very large station names, but keeps the preferred size and allows
full customizability.