applets/weather: don't use empty string as source
When no location has been set yet, the default value for the
weatherSource is an empty string. This is passed to the dataengine,
which complains about not finding an ion with that name. There does not
seem to be a way to get it to not use a value as the data source except
not setting one.

So we set the data source using a conditional binding that is only set
when the weatherSource is not the the empty string, and silence the
warning.