applets/notes: avoid scrollbar binding loop
The default PC3 Scrollbar's visibility binding causes a binding loop
for the notes applet. This is because the scrollbar becoming visible
reduces the available width and can so change the height of the word-
wrapped text, which then causes the visibility binding to be recomputed.
This would not be an endless loop, however – adding a scrollbar cannot
make the height shorter (so it'll always remain necessary if it was
necessary before showing the scroll bar), and there's at most one scroll
bar so it can't get even longer either. All it would take to resolve
this is a little time to let it settle.

So we give it a litle time with a delayed binding.