applets/weather: use list of QtObjects for model
The model that fills the information panels in the full representation
is built fresh using javascript code every time the data changes, which
causes the repeaters to recompute the contents as well. This is somewhat
wasteful, and is not very declarative.

Instead, we build a list of QtObjects with clearly specified types. This
allows us to have the property bindings that exist anyway do some of the
work of updating the data.