applets/fifteenPuzzle: only request image if there is one
Fifteen Puzzle always requests the pixmap for the image from the config,
even if there is no image set (i.e. the field is empty). This causes a
warning when the imageprovider can't actually provide the image that
doesn't exist.

Instead, only request the image if one is actually set.