applets/calculator: make arrow key navigation more predictable
The calculator applet allows arrow key navigation, but the base
representation is included in this (reachable from four different
buttons) where the user can't really do anything. For the other buttons.
moving outside the grid uses the representation's navigation bindings
as the default, which sometimes works but in others makes large
unpredictable jumps.

Instead, we explicitly set predictable orders for moving beyond the edge
of the field, using Pacman-style navigation (enters on the opposite side
of the same row or column).