applets/calculator: don't flip in RtL
Calculators seem to be customarily LtR even in RtL languages (likely
inherited from physical calculators). This is the case even when using
local numerals, but here even these are hardcoded to (Western) Arabic
numerals.

Instead, force LtR layout. The labels of buttons which may be translated
(e.g. Clear) are not affected.