applets/comic: drop unused configuration option
In the past, the Comic applet supported automatically checking for
provider upgrades from KNS, but this was removed in 2023 with d65893dc.
The configuration option in the UI is still there though, but does
nothing except create warnings when it can't load or save the option.

Instead, just get rid of it.