applets/comic: don't update comic when none are selected
The comic plugin regularly tries to load the current comic, and will
fail with a warning if it can't. This warning may sometimes be justified
(if there is an actual error) but is not necessary if it's just that no
comics are selected to be shown, the "Configure" button on the applet
should be enough to alert the user to the configuration. In particular,
the timer that checks every five minutes if it's a new day will print a
warning; with no comics selected this the timer doesn't need to run at
all.

Instead, check whether there are actually comics selected to be
displayed before trying to load a comic, or before setting the timer.