applets/comic: don't draw outside plasmoid
The conditional text elements can both end up outside the plasmoid; the
top because it can be long and does not elide, the bottom because the
item ends up with zero implicitWidth due to the somewhat circular
anchor setup. This also breaks the hyperlinks/jump-to-comic
functionality of the bottom text.

Instead, set elide and remove the unnecessary anchors.