applet/comic: also cache current comic unless checking for updates
The comic applet caches the comics it downloads so that it does not need
to keep re-downloading them. This cache works by the full id of the
comic, so the most recent comic is never cached. For various reasons,
the current comic is frequently accessed (even if another comic is
selected), meaning that many actions in the applet will have the side
effect of also (re-)downloading the current page of each enabled comic.
This causes unneccessary traffic and load on the comic's servers. But
we can't just cache the result, or checking for updates will no longer
work as it will just reuse the cached result instead of checking for new
ones.

Instead, we add a new member variable that controls whether the engine
should look in the cache for the current page. The update checker turns
this off when beginning the check, and turns it off once finished.
This way, the image will only be reloaded once per checking cycle,
rather than possibly many times.