applets/comic: avoid infinte download loop
Two slots are connected to the engine finishing; one being the update
checker, the other the setting of the current image. Both can themselves
trigger the engine to run again. This is particularly problematic if the
most recent  comic page is involved, as it has two names which bypasses
the protection against the infinite loop. In certain, but not uncommon,
situations, the two will trigger each other and keep downloading all the
comics forever.

Instead, only continue checking for new updates if there is an update in
progress, using the new method introduced in the previous commit. This
way, the current image for qml updating will not cause the update
checker to run again, avoiding the problem.