applets/calculator: remove clipboard dataengine and fix pasting
The calculator applet used the clipboard dataengine to inspect the
clipboard before pasting so that only numbers could be pasted. This
dataengine was removed a while ago, leaving only an error message when
starting Plasma or when actually trying to paste.

Instead, use an invisible TextEdit and paste there. This gets rid of the
warning and makes pasting work again.