applets/dict: replace OverlaySheet with Kirigami.Dialog
The HIG says OverlaySheet should only be used for readonly content, but
here we have an interactive listview.

This change switches it to a Kirigami.Dialog, and bases the general
interaction and design on the similar dialog in kcm_regionandlang. The
OK button is removed as the HIG recommends against using this as a
button label, and the semantically appropriate Close button would just
be a duplicate of the existing dialog close button, as this is an
instant apply dialog.