applets/comic: separate representations and use PlaceholderMessage
The comic applet defines switching sizes, but does not have compact and
full representations so it can't actually switch between anything. This
leads to an awkward ui in many ways - with no comics loaded, the panel
version is completely invisible, the dektop version can be resized so
its own configure button won't fit, with comics loaded the panel version
is just a tiny version of the comic with a rather awkward hover-based
popup UX that doesn't work on top panels, and is also non-functional if
more than one comic is selected.

Instead, have the current UI be a fullRepresentation and use the default
compactRepresentation. This makes the panel version work like a typical
applet - it's an icon, and if you click it you get a nicely interactive
popup fullRepresentation. The configuration helper message will not work
here, so we add a PlaceholderMessage, and for consistency and clarity we
also use it for the desktop version.