applets/comic: fix storing position
The store position functionality currently does not work for the first
(or only) tab; while the position is correctly set, when the qml loads
and sets the current tab, changeComic is called with parameter false
which ignores stored position and resets it.

Instead, just do nothing in this case - switching to the comic we're
already on doesn't really make sense in the first place.