applets/comic: use placeholderMessage for errors
The comic applet has some settings about displaying error images, but
these images don't seem to work, there doesn't seem to be code left that
actually implements them, and even if it worked this does not seem like
a modern Plasma approach. At present, if an error happens, the applet
either does not display anything, or confusingly displays the last image
that was loaded in another comic. Neither outcome seems particularly
satisfying, especially with a large number of very old comic plugins in
the store that no longer work.

This change passes the error state upward to qml, and adds a placeholder
message for these situations. If error display is turned off in the
settings, the behavior is unchanged for now.