applets/comic: only count each item once in cache
We can download some items multiple times, in particular the most recent
comic, as we need to check whether the strip has been replaced with a
new one, and that stores that comic both as the most recent and as its
permanent id. The cache does not handle this, and counts each time they
are downloaded toward the total cache limit; after the update checker
runs a couple of times all other entries will have been deleted from the
cache making it rather pointless.

Instead, we just remove previous entries with the same id from the list
of cached entries before saving them again. The files in the cache will
be overwritten (or already were, as the case may be), so there's no need
to handle them manually.