applets/mediaframe: fix pausing on hover
The widget tries to stop the timer when the user hovers over the image.
But this doesn't work - the mousearea in the fullRepresentation can't be
accessed by id (as it's a component). The only thing that happens is an
error message during loading, but no pausing.

Instead set up an alias in the fullRepresentation and access that
through the fullRepresentationItem.