applets/mediaframe: remove qml log messages
There are a few qml debug or warning messages left in the Media Frame
widget. These are either not particularly informative in the first place
(the applet already shows that no images are loaded) or redundant with
the log messages from the backend (which have more fine-grained control)
and if messages were necessary, they should be presented to the user
rather than dumped into the log.

This change removes these messages.