applets/timer: make seconds display work consistently
TimerEdit respects the setting on showing seconds. To do this, it
accesses the plasmoid configuration through the root element. The
component is also used in the configuration dialog to edit the list of
predefined timers, and this doesn't work there as it's not a descendant
of the plasmoid root.

Instead, make this a proper property of the TimerEdit, and set it as
appropriate from the point where it's used either through the root
PlasmoidItem or the cfg_ property.