applets/comic: set current also for cached comics
The current comic has special handling as the requested id does not
always match the one that plugins report (as some, but not all, return
the permanent id of the current comic). When the current comic is
retrieved from the cache, we don't set this, which breaks switching tabs
to such a comic.

Instead, set the current comic flag also for comics we retrieve from the
cache, This also allows us to make a complicated conditional a bit more
semantic.