applets/comic: store lastCachedStripIdentifier
The engine supports a metadata field called lastCachedStripIdentifier
which is saved and handled if a particular value is set, but it is never
explicitly set, so will always just be saved as an empty string.

Instead, store the latest successful result that was returned for a
request for the latest comic. This seems to be what was originally
intended and might sometimes do something useful by setting a valif
previous strip if there is an error.