applets/comic: allow accessing previous strip if current failed
Some comic strip plugins may fail on some days as they don't have an
image for that day (such as when APOD has a video instead of an image).
This currently breaks the applet so that even previous strips can no
longer be accessed, as the error display does not have the
navigation buttons available. And even if it did, while the engine
correctly sets the last cached strip (if set) as a potential previous
comic, the intermediate layer does not pass it to the qml if an error
is present.

Instead, we move the error placeholder into the comic view instead of
replacing it, and always pass the previous strip identifier to the
frontend if the engine sets one.