applets/notes: add button to keep panel popup pinned open
Tray popups and several other widgets like Kickoff have the ability to
keep themselves pinned open. The notes widget does not have this, but
this would be convenient when interacting with a window and a popup note
or when opening the clipboard popup.

This change adds such a button to the button row, next to the configure
button.

BUG: 501984