applets/comic: pass next/prev identifier if present even when error
The comic applet sets the previous identifier to the last cached one,
and the next identifier to nothing, in case there is an error. For
complete failure this is a good idea, but it's not a good experience if
there are less fatal errors. An example of this is the APOD plugin,
where the plugin will fail if the daily picture is a video instead of an
image, but the links to the previous/next images are still set correctly
where they are available (previous always, next for all except the
newest). In this case, attempting to go back from one such image will go
to the most recent (non-error) page rather than the previous day, and
going to the next image is not possible at all.

Instead, do not replace the next/previous identifiers if the plugin
reports them, and forward them to qml so the user can access them.