applets/comic: remove most recent cached comic when replaced
We store the most recent (in terms of release) comic both as the current
comic (with an unspecified comic identifier) and under its permanent id.
When a new strip is published, that one will take on the role of most
recent comic. But at this point, the cached permanent id of that comic
is no longer valid; it does not have a link to the next comic (as there
was no next comic when it was cached) so when the user presses the
"previous comic" button from the most recent one, they can't go forward
again.

Instead, we remove the permanent id of the previously most recent comic
from the cache when there is a newer one. The pre-fetcher should then
automatically grab it again when the current one is loaded. Just
updating the "nextIdentifier" field of the respective comic instead of
re-downloading the whole would break if several new comics were released
in between updates.