applets/comic: downgrade severity of plugin log messages
Comic plugins can use a print function to print debug and status output,
and some plugins make heavy use of that. Currently, each such message
leads to a warning-level log message. These messages are relevant to
plugin develpers, but generally not to end users - if there is an error,
the comic plugin will report that independently.

Instead, we change the severity for these messages to Info, and set the
default severity for printing messages to Warning. This way, they will
not be displayed by default, but plugin developers or users interested
in these messages for other reasons can still access them independently
of the comic applet's internal debug logging.

Along the way, we fix a typo in a log message.