applets/notes: set text property on buttons
The icon-only buttons do not have a text property set, and instead set
Accessible.name to make them accessible to screen readers. This is
generally not the preferred way to achieve an icon-only button.

Instead, set the text property (which automatically assigns
Accessible.name) and set the display property to IconOnly.