applets/comic: wrap fullRepresentation in Item
The previously used complex anchor setup was replaced with a simple
layout in c76ca0ac. This works fine for desktop widgets or panel popups,
but confuses the panel logic if the fullRepresentation is shown directly
in a very large panel, causing an endless loop if more than one thing is
visible at the same time (such as the main comic view and e.g. the tab
bar or the comic title label).

Instead, wrap the fullRepresentation in an Item, which solves the issue.

BUG: 502432