applets/comic: better handling of recent failure
The comic plugin stores the most recent failure and tries to reload it
when the network is online again. There are three issues with this:
first, the stored value is never cleared, so if there is a failue it
keep trying and trying; second, it will also try to download an empty
comic if there was no error, which can cause issues as the plugin does
not handle this always gracefully, and third, it's not quite clear why
the plugin does this in the general case: only the currently displayed
comic would need to be checked (as that one may be an error display that
is now fixed).

We solve all three issues by clearing the error when downloading a strip
succeeds and only requesting the failed identifier if it's not empty.