applets/timer: only prompt if timer list is changed
The list of predefined timers is stored as a (int)list. When entries are
added or removed, configurationChanged() is called, which marks the
configuration as changed until it is manually saved or discarded. As
entries are pushed/removed instead of recreating the list, it is
necessary to have some custom handling here.

Instead just call the property's Changed() signal and have the existing
configuration management logic handle this for us, prompting the user
only if the present state is different from the last saved one.