applets/comic: declare context menu in qml
Currently the context menu for the applet is defined in comic.cpp, which
leads to a rather imperative style and complex cpp applet code with lots
of action members that mixes frontend and backend concerns.

This change moves the action definitions into qml, where they can be
handled in a more declarative way, and works towards more clearly
separating the concerns. Unrelated changes are kept to a minimum, but
the "Show Actual Size" setting is now named that way throughout the code
(except for the config file key, which retains its existing name);
switching between three different names for the same thing was just too
confusing.