applets/dictionary: move Add button to header
The button for adding dictionaries to the list is currently below the
ListView. It feels a little lost there; usually we place such things in
the ListView header if there is one, or the page header if there isn't
(or in a separate Actions footer on mobile).

Instead, turn it into an Action and place it in the page's action
property, which will put it in an appropriate place.


Co-authored-by: Nate Graham <nate@kde.org>