applets/comic: move configuration out of applet
Currently, applet configuration is handled inside the applet manually by
storing the values in applet member variables/properties and populating/
setting them manually. Plasma's Applet class provides automatic handling
for this based on KConfigXT xml files and a fixed infrastructure for
configuration dialogs to set these values with various conveniences
(such as resonding to the current state automatically in many cases)
with much lower amounts of manual work.

This change switches to the xml-based version. Most of the configuration
values are only needed in the frontend, and the few that are required in
the backend can be read as needed