applets/comic: get rid of mDifferentComic and changeComic
mDifferentComic is a member variable that is only used once as an
argument to changeComic(), changeComic in turn only calls another method
depending on whether the argument is true (in all other cases it's
explicitly set to true. The one case where it can be false is when the
list of enabled comics changes; this case currently doesn't work very
well, but that is better handled on the qml side later, rather than with
complex state tracking in the applet's C++ code (that currently doesn't
really work anyway).

We can just get rid of the variable completely and immediately call the
function that changeComic would call to get rid of the unnecessary
indirection. Along the way, we get rid of a variable that is declared
and incremented a couple of times without ever being read.