applets/comic: get rid of sorted proxy model
The comic applet has three models for the comic list as members: the
original list as returned by KPackage (in model form), a
QSortFilterProxyModel that does nothing but sort the first model, and a
completely separate model that only contains the actually enabled comic
providers. The proxy model seems particularly unnecessary: the
underlying model does not change unless the package list is reloaded so
we can just sort it then; the unsorted list has no purpose.

This change does that and drops the proxy model.