applets/comic: store checked state in model
ComicModel has half an implementation for checked state, but half is
missing and the config dialog sets and reads the (Repeater) delegate
state manually. This is not optimal, as it locks us into using a
repeater; a ListView would be closer to our usual interface patterns but
as the delegates are loaded and unloaded they cannot store state.

This change completes the implementation os storing checked state in
ComicModel and switches the configuration dialog to use that.