applets/comic: remove updateContextMenu
With the context menu now being handled in QML, the only thing it does
(despite the name) is clear the active comics model if the current
identifier is empty. This does not seem to be a useful thing to do, as
it breaks the applet interface should it ever happen (which it shouldn't
in the first place).

Instead just remove it.