applets/comic: get rid of activeComicsModel
The applet has two underlying models, one containing all installed comic
providers, the other only the enabled ones (and it adds an extra role,
and renames some of the others away from standard rolenames). The second
model is only used at one point, namely to populate the list of tabs.

Instead, we add the necessary information (tab highlightes state and
enabledness) to the first model, and just use a KSortFilterProxyModel to
filter it down to populate the tabbar. This reduces complexity by having
fewer moving parts and simpler code paths.