applets/comic: don't immediately check on day change
The applet immediately updates the current comic when the date changes.
This doesn't seem to be warranted; given that users and comics are
international, the release schedule for even a date-based provider like
APoD is likely to not align so that the new image is available
immediately at midnight.

Instead, just drop that. The 5-minute timer is kept for now to handle
the case where the image is not available, and renamed as such.