applets/notes: don't set Layout preferred sizes
Currently, the fullRepresentation sets Layout.preferredSize. This gives
the widget a default size for the desktop representation or the panel
popup, but is much too eager on desktop - when the desktop layout is
recomputed (such as on plasmashell startup), it will try to give the
widget that size. If space for the preffered size is available in either
dimension, the widget size that the user chose will be overwritten by
the default size. Messing with the user's desktop widget arrangement in
this way is not nice; we should respect the user's choices where
possible.

Instead, set the width and height of the PlasmoidItem. A newly added
widget will use this for its default size, but will not override any
sizes the user chooses later.

BUG: 495306
FIXED-IN: 6.3.5


(cherry picked from commit 4b4090d1a238f6f5c3aaae4e9838b02342c22b61)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>