applets/comic: use configurationRequired for tiny widgets
Since the original implementation of resizing the widget on showing a
PlaceholderMessage was abandonded, the placeholder could escape the area
of the widget. This looks broken.

Instead, if the placeholderMessage is too large to fit, fall back to the
default Plasma widget method for indicating that configuration is
required: a simple button labeled "Configure".